Python/requirements.txt
pycryptodome
tqdm
